# docs/OPENAPI_FULL.yaml
openapi: 3.0.3
info:
  title: SMIF Gestão API — Full
  version: 0.3.0
  description: Especificação completa (Auth, Admin, RBAC, Localities, Specialties,
    Meetings, Tasks, Reports, Notices, Checklists, KPIs, Org/Elos, Dashboards).
servers:
- url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        details: {}
      required:
      - message
...
    TaskTemplate:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        phaseId:
          type: string
        specialtyId:
          type: string
          nullable: true
        appliesToAllLocalities:
          type: boolean
        reportRequiredDefault:
          type: boolean
      required:
      - id
      - title
      - phaseId
      - appliesToAllLocalities
      - reportRequiredDefault
...
security:
- bearerAuth: []
paths:
...
  /task-templates:
    get:
      tags:
      - Tasks
      responses:
        '200':
          description: OK
    post:
      tags:
      - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskTemplate'
      responses:
        '201':
          description: Created
  /task-templates/{id}/generate-instances:
    post:
      tags:
      - Tasks
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Created
...
  /task-instances/{id}/status:
    put:
      tags:
      - Tasks
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
              required:
              - status
      responses:
        '200':
          description: OK
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /task-instances/{id}/progress:
    put:
      tags:
      - Tasks
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                progressPercent:
                  type: integer
              required:
              - progressPercent
      responses:
        '200':
          description: OK
  /task-instances/{id}/assign:
    put:
      tags:
      - Tasks
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                assignedToId:
                  type: string
                  nullable: true
              required:
              - assignedToId
      responses:
        '200':
          description: OK
...
  /task-instances/calendar:
    get:
      tags:
      - Calendar
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
...